day 9 in ruby...
the circle node that i created is basically a double linked list, except
the first node initially links back to itself, rather than pointing to
nils.

working with arrays in ruby is just... very expensive when the amount of
data is large and insertion/deletion occurs in the middle of the array
(ie. copy all data forward or backwards 1 spot). double linked list
allows O(1) inserts and deletes (as long as you have the node you're
inserting after or deleting.